\name{sim.idmModel}
\alias{sim.idmModel}
\title{Simulate illness-death model data}
\usage{
\method{sim}{idmModel}(x, n, illness.known.at.death = TRUE, compliance = 1,
  latent = FALSE, keep.inspectiontimes = FALSE, ...)
}
\arguments{
  \item{x}{An \code{idmModel} object as obtained with
  \code{idmModel}}

  \item{n}{Number of observations}

  \item{illness.known.at.death}{Affects the value of
  variable seen.ill}

  \item{compliance}{Probability of missing an inspection
  time.}

  \item{latent}{if TRUE keep the latent event times}

  \item{keep.inspectiontimes}{if \code{TRUE} keep the
  inspection times.}

  \item{...}{Extra arguments given to \code{sim}}
}
\value{
A data set with interval censored observations from an
illness-death model
}
\description{
Function to simulate illness-death model data
}
\details{
Based on the functionality of the lava PACKAGE
}
\examples{
example(idmModel)
help(idmModel)
}
\author{
Thomas Alexander Gerds
}

