% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SmoothHazard-package.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Data set for survival models: right-censored and interval-censored data.}
\format{A data frame with 936 observations on the following 4 variables.
\describe{ \item{l}{for diseased subjects: left endpoint of
censoring interval; for non-diseased subjects: right censoring time}
\item{r}{for diseased subjects: right endpoint of censoring
interval; for non-diseased subjects: right censoring time for the disease
event} \item{id}{disease status} \item{cov}{covariate} }}
\description{
A simulated data frame for survival models composed of right-censored and
interval-censored data.
}
\examples{
data(testdata)
head(testdata)
}
\keyword{datasets}

