% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SmoothHazard-package.R
\docType{data}
\name{Paq1000}
\alias{Paq1000}
\title{Paquid data set}
\format{A data frame with 1000 rows and the following 8 columns.  \describe{
\item{dementia}{dementia status, 0=non-demented, 1=demented}
\item{death}{death status, 0=alive, 1=dead} \item{e}{age at
entry in the study} \item{l}{for demented subjects: age at the visit
before the diagnostic visit; for non-demented subjects: age at the last
visit (censoring age)} \item{r}{for demented subjects: age at the
diagnostic visit; for non-demented subjects: age at the last visit
(censoring age)} \item{t}{for dead subjects: age at death; for alive
subject: age at the latest news} \item{certif}{primary school
certificate:\code{0=with certificate}, \code{1=without certificate}}
\item{gender}{gender: \code{0=female}, \code{1=male}} }}
\description{
Paquid data set composed of 1000 subjects selected randomly from the Paquid
data set of 3675 subjects.
}
\examples{
data(Paq1000)
}
\keyword{datasets}

