% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.idm.R
\name{predict.idm}
\alias{predict.idm}
\title{Predictions for an illness-death model using either a penalized likelihood
approach or a Weibull parametrization.}
\usage{
\method{predict}{idm}(object, s, t, newdata, nsim = 200, seed = 21,
  conf.int = TRUE, level = 0.95, lifeExpect = FALSE, ...)
}
\arguments{
\item{object}{an \code{idm} class objects returned by a call to the
\code{\link{idm}} function}

\item{s}{time point at which prediction is made.}

\item{t}{time horizon for prediction.}

\item{newdata}{A data frame with covariate values for prediction.}

\item{nsim}{number of simulations for the confidence intervals
calculations.  The default is 200.}

\item{seed}{Seed passed to \code{set.seed} for Monte Carlo
simulation of confidence intervals.}

\item{conf.int}{Logical: with (\code{TRUE}) or without
(\code{FALSE}) confidence intervals for the predicted values. The
default is \code{TRUE}.}

\item{level}{Value between 0 and 1, the level of confidence. Default is .95.}

\item{lifeExpect}{Logical. If \code{TRUE} compute life
expectancies, i.e., \code{t=Inf}.}

\item{...}{other parameters.}
}
\value{
a list containing the following predictions with pointwise
confidence intervals: \item{p00}{the transition probability \eqn{p_{00}}.}
\item{p01}{the transition probability \eqn{p_{01}}.} \item{p11}{the
transition probability \eqn{p_{11}}.} \item{p12}{the transition probability
\eqn{p_{12}}.} \item{p02_0}{the probability of direct transition from state
0 to state 2.} \item{p02_1}{the probability of transition from state 0 to
state 2 via state 1.} \item{p02}{transition probability \eqn{p_{02}}. Note
that \code{p02}=\code{p_02_0}+\code{p02_1}.} \item{F01}{the lifetime risk of
disease. \code{F01}=\code{p01}+\code{p02_1}.} \item{F0.}{the probability of
exit from state 0. \code{F0.}=\code{p02_0}+\code{p01}+\code{p02_1}.}
}
\description{
Predict transition probabilities and cumulative probabilities from an object
of class \code{idmSplines} with confidence intervals are calculated.
}
\examples{
\dontrun{
d=simulateIDM(n = 100)
fit <- idm(formula01=Hist(time=list(L,R),event=seen.ill)~X1+X2+X3,
               formula02=Hist(time=observed.lifetime,event=seen.exit)~X1+X2+X3,
               data=d,conf.int=FALSE)
predict(fit,s=0,t=80,nsim=3,conf.int=FALSE,lifeExpect=F)


data(Paq1000)
library(prodlim)
fit <- idm(formula02=Hist(time=t,event=death,entry=e)~certif,
formula01=Hist(time=list(l,r),event=dementia)~certif,data=Paq1000)

predict(fit,s=70,t=80,newdata=data.frame(certif=1))
predict(fit,s=70,lifeExpect=TRUE,t=80,newdata=data.frame(certif=1))

fit.splines <-  idm(formula02=Hist(time=t,event=death,entry=e)~certif,
		formula01=Hist(time=list(l,r),event=dementia)~certif,
                formula12=~1,
                method="Splines",
		data=Paq1000)

pred <- predict(fit.splines,s=70,t=80,Z01=c(1),Z02=c(1))
pred

}
}
\author{
R: Celia Touraine <Celia.Touraine@isped.u-bordeaux2.fr> and Thomas Alexander Gerds <tag@biostat.ku.dk> Fortran:
Pierre Joly <Pierre.Joly@isped.u-bordeaux2.fr>
}
\seealso{
\code{\link{idm}}
}
\keyword{methods}

