\name{simulateIDM}
\alias{simulateIDM}
\title{Sample illness-death model data}
\usage{
simulateIDM(n = 100)
}
\arguments{
  \item{n}{number of observations}
}
\description{
Function to simulate from a specific illness-death model
}
\details{
For the purpose of illustrating the help pages of the
SmoothHazard package.
}

